feat: redesign registration section - add images to the section - refactor code
